+++ src/lib3to6/__init__.py
@@ -7,8 +7,9 @@
 from .utils import parsedump_ast
 from .utils import parsedump_source
 from .packaging import fix
+from .packaging import build_py
 from .transpile import transpile_module
 
 __version__ = "v202107.1047"
 
-__all__ = ["fix", "transpile_module", "parsedump_ast", 'parsedump_source']
+__all__ = ["fix", "transpile_module", "parsedump_ast", 'parsedump_source', 'build_py']