+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
  - Add better default integration method: [`lib3to6.build_py`][href_readme_integration]
  - Add new default integration method: [`lib3to6.build_py`][href_readme_integration]
+ - Deprecate old integration method (had issues with package data).
  - Fix invalid parsing of install_requires.
  - Fix cache invalidation when build config changes.
  - Fix type annotations with constants (e.g. `Optional['MyClass']`)